Rename file with soft link class
